+++ wwwroot/ProjList.aspx.cs
@@ -44,11 +44,13 @@ public partial class _Default : System.Web.UI.Page
          try
          {
              con.Open();
+             
              SqlDataReader reader = cmd.ExecuteReader();
              DataTable dt = new DataTable();
              dt.Load(reader);
              ProjListGrid.DataSource = dt;
              ProjListGrid.DataBind();
+            
 
          }
          finally
@@ -107,10 +109,32 @@ public partial class _Default : System.Web.UI.Page
 
         string pid = ProjListGrid.DataKeys[e.RowIndex].Values["PID"].ToString();
         con.Open();
-        SqlCommand cmd = new SqlCommand("delete from Project where pid='" + pid + "'", con);
-        int result = cmd.ExecuteNonQuery();
-        con.Close();
-        getAllProj();
+
+        AdminMaster master = (AdminMaster)Page.Master;
+//---------------------------------        
+        SqlCommand cmd0 = new SqlCommand("select count(*) from Assignment where  PID ='"+pid+"'",con);
+        int count = (int)cmd0.ExecuteScalar();
+        if (count != 0)
+        {//uname exisits                
+            master.AlertWarning("Project already existed in Assignment, Please delete assignment first!");
+        }
+        else 
+        {
+            SqlCommand cmd1 = new SqlCommand("select count(*) from student where  PID ='"+pid+"'",con);
+            int count1 = (int)cmd1.ExecuteScalar();
+            if (count1 != 0)
+            {
+                master.AlertWarning("Project already existed in Student, Please delete assignment first!");
+            }
+            else
+            {
+                SqlCommand cmd = new SqlCommand("delete from Project where pid='" + pid + "'", con);
+                int result = cmd.ExecuteNonQuery();
+                con.Close();
+                getAllProj();
+                master.AlertWarning("Deleted successfully");
+            }
+        }
     }
     protected void ProjListGrid_RowEditing(object sender, GridViewEditEventArgs e)
     {
@@ -121,11 +145,13 @@ public partial class _Default : System.Web.UI.Page
     {
         string pid = ProjListGrid.DataKeys[e.RowIndex].Values["PID"].ToString();
         TextBox pn = (TextBox)ProjListGrid.Rows[e.RowIndex].FindControl("textbox1");
-        TextBox cid = (TextBox)ProjListGrid.Rows[e.RowIndex].FindControl("textbox2");
+        DropDownList aa = (DropDownList)ProjListGrid.Rows[e.RowIndex].FindControl("CA");
+        string cid = aa.SelectedItem.Value;
+        //TextBox cid = (TextBox)ProjListGrid.Rows[e.RowIndex].FindControl("textbox2");
         TextBox fn = (TextBox)ProjListGrid.Rows[e.RowIndex].FindControl("textbox3");
         TextBox gid = (TextBox)ProjListGrid.Rows[e.RowIndex].FindControl("textbox4");
         con.Open();
-        SqlCommand cmd = new SqlCommand("update Project set PName='" + pn.Text + "',cid='" + cid.Text + "',gid='" + gid.Text + "' where pid='" + pid + "'", con);
+        SqlCommand cmd = new SqlCommand("update Project set PName='" + pn.Text + "',cid='" + cid + "',gid='" + gid.Text + "' where pid='" + pid + "'", con);
         SqlCommand cmd1 = new SqlCommand("update student set FName='" + fn.Text + "' where pid='" + pid + "'", con);
         int result = cmd.ExecuteNonQuery();
         int result1 = cmd1.ExecuteNonQuery();
@@ -134,4 +160,9 @@ public partial class _Default : System.Web.UI.Page
         getAllProj();
 
     }
+
+    protected void ProjListGrid_DataBound(object sender, EventArgs e) 
+    {         
+    }
+
 }
